Actually working variant of previous commit
You know, there are some days where you are reflecting on why you even
write this stuff. This is one of these days and today I realized that a
theoretical long term goal might be that this becomes actually usable.
Based on that I decided that committing actually correct code might be a
pretty good idea.